fix compensation spillover problems
Some versions of .fcs files can save the spillover matrix for compensation into different matrix positions. Prior this, it was using the first matrix but for fcs 2.0 it was the third matrix. Now, it raises an error if there is no compensation matrix, checks which ones are not empty and uses the first non-empty matrix.